+++ CHANGELOG.md
@@ -67,7 +67,9 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
+* **about.jsx:** add y3-shadow-md-inner class to About page ([4d82459](https://github.com/chingu-voyages/v46-tier2-team-19/commit/4d82459a4c483bcb430ba9ac814842cdfed1c1b2))
 * **api:** add new tips.js file to fetch tips from Tasty API ([da69610](https://github.com/chingu-voyages/v46-tier2-team-19/commit/da69610c0c2e18c5daabfc02521618f16e05a996))
+* **global.css:** create y3-shadow-md-inner class ([7c67434](https://github.com/chingu-voyages/v46-tier2-team-19/commit/7c674343fbcc3be4d6c6a3dd24bff81b3db03f28))
 * **mirageserver:** add tips.json and endpoint for tips in server.js ([ad337dc](https://github.com/chingu-voyages/v46-tier2-team-19/commit/ad337dcc0b65abd43131ac8c80b571a569d583c0))
 * **package.json, package-lock.json:** add react-responsive-carousel package ([713c339](https://github.com/chingu-voyages/v46-tier2-team-19/commit/713c3395219ec7b41c6cd31afb7c76b99af17ca4))
 * **recipe details:** continue improving responsive layout ([83e4357](https://github.com/chingu-voyages/v46-tier2-team-19/commit/83e4357964e724c1919ca2e3a32f26c4ebffaa24))
@@ -79,6 +81,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 * **about.jsx:** fix responsive spacing around team members at larger viewport widths ([6b2c4ee](https://github.com/chingu-voyages/v46-tier2-team-19/commit/6b2c4ee830ee494ff6e426a80467ba9ce6ab9ffe))
 * **about.jsx:** remove stroke from team member role heading ([d50ab83](https://github.com/chingu-voyages/v46-tier2-team-19/commit/d50ab83bee449097aab00061dabd3dd383908c03))
+* **about.jsx:** fix YumYumTeam heading responsiveness ([b38bcf3](https://github.com/chingu-voyages/v46-tier2-team-19/commit/b38bcf36aa32ee30aa925c0c414600d534ce1b1c))
 * **fetchrecipes:** wrap localStorage.setItem in try/catch block ([14330bc](https://github.com/chingu-voyages/v46-tier2-team-19/commit/14330bcfd6a2a7879f9c40d4e4bd62c5e79f1dda))
 * **matchsearchid:** replace negative lookahead with capturing group (for Safari) ([9a79455](https://github.com/chingu-voyages/v46-tier2-team-19/commit/9a79455099d8d0c2f6af5902bb78f858a06572d8))
 
@@ -99,6 +102,8 @@ All notable changes to this project will be documented in this file. See [standa
 
 * **about.jsx:** fix responsive spacing around team members at larger viewport widths ([6b2c4ee](https://github.com/chingu-voyages/v46-tier2-team-19/commit/6b2c4ee830ee494ff6e426a80467ba9ce6ab9ffe))
 * **about.jsx:** remove stroke from team member role heading ([d50ab83](https://github.com/chingu-voyages/v46-tier2-team-19/commit/d50ab83bee449097aab00061dabd3dd383908c03))
+* **about.jsx:** fix YumYumTeam heading responsiveness ([b38bcf3](https://github.com/chingu-voyages/v46-tier2-team-19/commit/b38bcf36aa32ee30aa925c0c414600d534ce1b1c))
+* **about.jsx:** removed white stroke around main heading & deleted deprecated comments ([9453c1e](https://github.com/chingu-voyages/v46-tier2-team-19/commit/9453c1e545b666b7b97a71fa8d8f83e899124f85))
 * **fetchrecipes:** wrap localStorage.setItem in try/catch block ([14330bc](https://github.com/chingu-voyages/v46-tier2-team-19/commit/14330bcfd6a2a7879f9c40d4e4bd62c5e79f1dda))
 * **matchsearchid:** replace negative lookahead with capturing group (for Safari) ([9a79455](https://github.com/chingu-voyages/v46-tier2-team-19/commit/9a79455099d8d0c2f6af5902bb78f858a06572d8))